requirements.txt
yfinance
numpy
pandas
google-cloud-bigquery
pydantic-settings
pyyaml
anthropic
ruff
pytest
